STITCH-778 - v2 login for admin
move admin to its own file
logout, userProfile, doSessionPost functions added to admin.js
doSessionPost added to client.js
