+++ src/common.js
@@ -9,6 +9,8 @@ export const IMPERSONATION_ACTIVE_KEY = '_stitch_impers_active';
 export const IMPERSONATION_USER_KEY = '_stitch_impers_user';
 export const IMPERSONATION_REAL_USER_AUTH_KEY = '_stitch_impers_real_ua';
 export const USER_AUTH_COOKIE_NAME = 'stitch_ua';
+export const CLIENT_TYPE = 'client';
+export const ADMIN_TYPE = 'admin';
 export const DEFAULT_STITCH_SERVER_URL = 'https://stitch.mongodb.com';
 
 // VERSION is substituted with the package.json version number at build time
