+++ test/testutil.js
@@ -3,6 +3,6 @@ const constants = require('./constants');
 
 export async function getAuthenticatedClient(apiKey, serverUrl) {
   const adminClient = new stitch.Admin(serverUrl || constants.DEFAULT_SERVER_URL);
-  await adminClient.client.authenticate('apiKey', apiKey);
+  await adminClient.authenticate('apiKey', apiKey);
   return adminClient;
 }
